requirements.txt
asgiref==3.8.1
Django==5.1.6
django-filter==25.1
django-querycount==0.8.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Markdown==3.7
PyJWT==2.10.1
sqlparse==0.5.3
